Move the SPI interfaces to .spi.* packages
