+++ api/src/main/java/io/opentelemetry/spi/metrics/package-info.java
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/** The SPI (service provider interface) for initializing the OpenTelemetry tracing API. */
+/** The SPI (service provider interface) for initializing the OpenTelemetry metrics API. */
 @ParametersAreNonnullByDefault
-package io.opentelemetry.api.trace.spi;
+package io.opentelemetry.spi.metrics;
 
 import javax.annotation.ParametersAreNonnullByDefault;
