+++ sdk/metrics/src/main/resources/META-INF/services/io.opentelemetry.spi.metrics.MeterProviderFactory
