+++ sdk/tracing/src/main/java/io/opentelemetry/sdk/trace/spi/TracerProviderFactorySdk.java
@@ -6,8 +6,8 @@
 package io.opentelemetry.sdk.trace.spi;
 
 import io.opentelemetry.api.trace.TracerProvider;
-import io.opentelemetry.api.trace.spi.TracerProviderFactory;
 import io.opentelemetry.sdk.trace.TracerSdkProvider;
+import io.opentelemetry.spi.trace.TracerProviderFactory;
 
 /** SDK implementation of the TracerProviderFactory for SPI. */
 public final class TracerProviderFactorySdk implements TracerProviderFactory {
